refactor away PlayerInteractionViewModel
