+++ Screenbox.Core/Models/PositionChangedResult.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace Screenbox.Core.Models;
+
+public record struct PositionChangedResult(TimeSpan OldPosition, TimeSpan NewPosition, TimeSpan OriginalPosition, TimeSpan NaturalDuration)
+{
+}
