+++ Screenbox.Core/ViewModels/PlayerPageViewModel.cs
@@ -7,7 +7,9 @@ using CommunityToolkit.Mvvm.Messaging.Messages;
 using Microsoft.Toolkit.Uwp.UI;
 using Screenbox.Core.Enums;
 using Screenbox.Core.Events;
+using Screenbox.Core.Helpers;
 using Screenbox.Core.Messages;
+using Screenbox.Core.Models;
 using Screenbox.Core.Playback;
 using Screenbox.Core.Services;
 using System;
@@ -46,6 +48,10 @@ namespace Screenbox.Core.ViewModels
         [ObservableProperty]
         [NotifyPropertyChangedRecipients]
         private PlayerVisibilityState _playerVisibility;
+
+        [ObservableProperty]
+        [NotifyPropertyChangedRecipients]
+        private MediaPlaybackState _playbackState;
 
         public bool SeekBarPointerInteracting { get; set; }
 
@@ -200,7 +206,7 @@ namespace Screenbox.Core.ViewModels
         public void OnVolumeKeyboardAcceleratorInvoked(KeyboardAccelerator sender, KeyboardAcceleratorInvokedEventArgs args)
         {
             if (_mediaPlayer == null || sender.Modifiers != VirtualKeyModifiers.None) return;
-            args.Handled = true;
+            bool playerVisible = PlayerVisibility == PlayerVisibilityState.Visible;
             int volumeChange;
             VirtualKey key = sender.Key;
 
@@ -208,19 +214,159 @@ namespace Screenbox.Core.ViewModels
             {
                 case (VirtualKey)0xBB:  // Plus ("+")
                 case (VirtualKey)0x6B:  // Add ("+")(Numpad plus)
+                case VirtualKey.Up when playerVisible:
                     volumeChange = 5;
                     break;
                 case (VirtualKey)0xBD:  // Minus ("-")
                 case (VirtualKey)0x6D:  // Subtract ("-")(Numpad minus)
+                case VirtualKey.Down when playerVisible:
                     volumeChange = -5;
                     break;
                 default:
-                    args.Handled = false;
                     return;
             }
 
             int volume = Messenger.Send(new ChangeVolumeRequestMessage(volumeChange, true));
             Messenger.Send(new UpdateVolumeStatusMessage(volume, false));
+            args.Handled = true;
+        }
+
+        public void OnSeekKeyboardAcceleratorInvoked(KeyboardAccelerator sender,
+            KeyboardAcceleratorInvokedEventArgs args)
+        {
+            if (_mediaPlayer == null) return;
+            bool playerVisible = PlayerVisibility == PlayerVisibilityState.Visible;
+            long seekAmount = 0;
+            int direction;
+            switch (sender.Key)
+            {
+                case VirtualKey.Left when playerVisible:
+                case VirtualKey.J:
+                    direction = -1;
+                    break;
+                case VirtualKey.Right when playerVisible:
+                case VirtualKey.L:
+                    direction = 1;
+                    break;
+                default:
+                    return;
+            }
+
+            switch (sender.Modifiers)
+            {
+                case VirtualKeyModifiers.Control:
+                    seekAmount = 10000;
+                    break;
+                case VirtualKeyModifiers.Shift:
+                    seekAmount = 1000;
+                    break;
+                case VirtualKeyModifiers.None:
+                    seekAmount = 5000;
+                    break;
+            }
+
+            seekAmount *= direction;
+            if (seekAmount != 0)
+            {
+                Messenger.SendSeekWithStatus(TimeSpan.FromMilliseconds(seekAmount));
+            }
+
+            args.Handled = true;
+        }
+
+        public void OnPercentJumpKeyboardAcceleratorInvoked(KeyboardAccelerator sender,
+            KeyboardAcceleratorInvokedEventArgs args)
+        {
+            if (_mediaPlayer == null || PlayerVisibility != PlayerVisibilityState.Visible) return;
+            VirtualKey key = sender.Key;
+            switch (key)
+            {
+                case VirtualKey.NumberPad0:
+                case VirtualKey.NumberPad1:
+                case VirtualKey.NumberPad2:
+                case VirtualKey.NumberPad3:
+                case VirtualKey.NumberPad4:
+                case VirtualKey.NumberPad5:
+                case VirtualKey.NumberPad6:
+                case VirtualKey.NumberPad7:
+                case VirtualKey.NumberPad8:
+                case VirtualKey.NumberPad9:
+                    int percent = (key - VirtualKey.NumberPad0) * 10;
+                    TimeSpan newPosition = _mediaPlayer.NaturalDuration * (0.01 * percent);
+                    PositionChangedResult result = Messenger.Send(new ChangeTimeRequestMessage(newPosition));
+                    newPosition = result.NewPosition;
+                    Messenger.SendPositionStatus(newPosition, result.NaturalDuration, $"{percent}%");
+                    break;
+                default:
+                    return;
+            }
+
+            args.Handled = true;
+        }
+
+        public void OnPlaybackRateKeyboardAcceleratorInvoked(KeyboardAccelerator sender,
+            KeyboardAcceleratorInvokedEventArgs args)
+        {
+            if (_mediaPlayer == null || sender.Modifiers != VirtualKeyModifiers.Shift ||
+                PlayerVisibility != PlayerVisibilityState.Visible) return;
+            args.Handled = true;
+            switch (sender.Key)
+            {
+                case (VirtualKey)190:   // Shift + . (">")
+                    TogglePlaybackRate(true);
+                    return;
+                case (VirtualKey)188:   // Shift + , ("<")
+                    TogglePlaybackRate(false);
+                    return;
+            }
+        }
+
+        public void OnFrameJumpKeyboardAcceleratorInvoked(KeyboardAccelerator sender,
+            KeyboardAcceleratorInvokedEventArgs args)
+        {
+            if (PlayerVisibility != PlayerVisibilityState.Visible || (!(_mediaPlayer?.CanSeek ?? false)) ||
+                _mediaPlayer.PlaybackState != MediaPlaybackState.Paused) return;
+            args.Handled = true;
+            switch (sender.Key)
+            {
+                case (VirtualKey)190:   // Period (".")
+                    _mediaPlayer.StepForwardOneFrame();
+                    return;
+                case (VirtualKey)188:   // Comma (",")
+                    _mediaPlayer.StepBackwardOneFrame();
+                    return;
+            }
+        }
+
+        private void TogglePlaybackRate(bool speedUp)
+        {
+            if (_mediaPlayer == null) return;
+            Span<double> steps = stackalloc[] { 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2 };
+            double lastPositiveStep = steps[0];
+            foreach (double step in steps)
+            {
+                double diff = step - _mediaPlayer.PlaybackRate;
+                if (speedUp && diff > 0)
+                {
+                    _mediaPlayer.PlaybackRate = step;
+                    Messenger.Send(new UpdateStatusMessage($"{step}×"));
+                    return;
+                }
+
+                if (!speedUp)
+                {
+                    if (-diff > 0)
+                    {
+                        lastPositiveStep = step;
+                    }
+                    else
+                    {
+                        _mediaPlayer.PlaybackRate = lastPositiveStep;
+                        Messenger.Send(new UpdateStatusMessage($"{lastPositiveStep}×"));
+                        return;
+                    }
+                }
+            }
         }
 
         partial void OnPlayerVisibilityChanged(PlayerVisibilityState value)
@@ -338,6 +484,7 @@ namespace Screenbox.Core.ViewModels
 
             _dispatcherQueue.TryEnqueue(() =>
             {
+                PlaybackState = state;
                 IsPlaying = state == MediaPlaybackState.Playing;
                 IsOpening = false;
